Change name and role of NestedArgument
