--- src/main/java/me/s3ns3iw00/jcommands/argument/NestedArgument.java
@@ -1,9 +1,0 @@
-package me.s3ns3iw00.jcommands.argument;
-
-/**
- * A helper class
- * It helps to identify arguments that can be contained by {@link me.s3ns3iw00.jcommands.argument.type.ConstantArgument}
- * Arguments that extends this class can be nested
- */
-public abstract class NestedArgument extends Argument {
-}
